move CURRENT_PLATFORM + add get_current_platform
was in: wakepy.core.configuration
now in: wakepy.core.platform

Added function which gets the current platform. It is one of WINDOWS,
LINUX, MACOS and OTHER.
