+++ wakepy/core/__init__.py
@@ -7,12 +7,12 @@ See the public Python API at: https://wakepy.readthedocs.io/
 from .activationresult import ActivationResult as ActivationResult
 from .calls import Call as Call
 from .calls import DbusMethodCall as DbusMethodCall
-from .configuration import CURRENT_PLATFORM as CURRENT_PLATFORM
 from .constants import BusType as BusType
 from .constants import ModeName as ModeName
 from .constants import PlatformName as PlatformName
+from .dbus import DbusAdapter as DbusAdapter
 from .dbus import DbusAddress as DbusAddress
 from .dbus import DbusMethod as DbusMethod
-from .dbus import DbusAdapter as DbusAdapter
 from .method import Method as Method
+from .platform import CURRENT_PLATFORM as CURRENT_PLATFORM
 from .strenum import StrEnum as StrEnum
